+++ src/main/java/roomescape/controller/ReservationController.java
@@ -13,9 +13,9 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.LoginUserEmail;
 import roomescape.auth.UserEmail;
 import roomescape.config.TokenPropertiesConfig;
-import roomescape.domain.reservation.ReservationService;
-import roomescape.domain.reservationtime.ReservationTimeService;
-import roomescape.domain.theme.ThemeService;
+import roomescape.service.ReservationService;
+import roomescape.service.ReservationTimeService;
+import roomescape.service.ThemeService;
 import roomescape.dto.ReservationAddRequestDto;
 import roomescape.dto.ReservationResponseDto;
 import roomescape.entities.Reservation;
