+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.reservationtime.ReservationTimeService;
+import roomescape.service.ReservationTimeService;
 import roomescape.dto.AvailableTimeResponse;
 import roomescape.dto.ReservationTimeAddRequestDto;
 import roomescape.entities.ReservationTime;
