+++ src/main/java/roomescape/service/MemberService.java
@@ -1,4 +1,4 @@
-package roomescape.domain.member;
+package roomescape.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
