+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservationtime;
+package roomescape.service;
 
 
 import org.springframework.stereotype.Service;
