+++ src/main/java/roomescape/service/ThemeService.java
@@ -1,4 +1,4 @@
-package roomescape.domain.theme;
+package roomescape.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
